Modelo 5/Clase 7 y 8 - Proyecto de integración/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rich==13.9.2
email-validator
